Add files via upload
The project demonstrates how wearable technology may be used beyond the hype to improve personal lifestyles by suggesting areas where generating a change in daily or routine habits may lead to a healthier lifestyle. Currently, such cycles appear to be driven by irrelevant or incomplete information which may not be easy to interpret but under peer pressure and the fear of exclusion, wearers adopt ad hoc change regardless. My project demonstrates that a planned approach to data collection, analysis and visualization coupled with a machine learning recommender approach may transpose wearer decisions to be informed and that the use of peer group data is for generating correlated recommendations not for applying social pressure.
The results also reveal that males are more likely to be overweight and fall into the sedentary range than females are. Unlike figures do not correlate sleep and daily activity. Sundays the least active with walking being the most popular activity. Daily tracking of sleep efficiency has picked up that this rises during weekdays but declines at weekends despite a higher sleep activity during weekends